pseudo 1.6.5: less pointlessly chatty
There was a stupid logic error controlling the diagnostic for
a "possible" mismatch involving trailing slashes and whether or
not a node was believed to be a directory. Specifically, a diagnostic
got printed any time a lookup for a directory *didn't* have a
trailing slash, as well as in the (actually intended) case where
a non-directory lookup *did*.

No other changes, but that one is probably significant.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Werner Grift <sky.captin@gmail.com>